pyproject.toml
[build-system]
requires = ["setuptools >= 75.3.2"]
build-backend = "setuptools.build_meta"

[project]
name = "Lokus"
version = "1.0.0"
description = "Lokus: Find issues in your APIs from the docs"
authors = [
    {name = "Vitor Cardoso", email = "vitor.cardoso5@fatec.sp.gov.br"},
    {name = "Robson Damasceno"},
    {name = "Virgílio Oliveira"}
]
maintainers = [
    {name = "Vitor Cardoso", email = "vitor.cardoso5@fatec.sp.gov.br"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.8",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "reportlab>=4.4.1",
]
keywords = ["security", "lgpd", "swagger", "api", "openapi"]

[dependency-groups]
dev = [
    "pre-commit>=3.5.0",
    "pytest>=8.3.5",
]


[project.scripts]
swagger-validator = "lokus.cli:main"

[tool.setuptools]
packages = ["lokus"]
